requirements.txt
#streamlit==1.39.0
#openai>=1.40.0
#python-dotenv==1.0.1
#SQLAlchemy==2.0.36
#pydantic==2.9.2
#faiss-cpu==1.8.0.post1
#tiktoken==0.7.0
#pypdf==5.0.0
#email-validator==2.2.0
#supabase

streamlit
google-generativeai>=0.7.2
supabase
faiss-cpu
pypdf
tiktoken
numpy
python-dotenv
email-validator
sentence-transformers
fpdf2